Refactor for UI split/other D20s, spellcasting, some class work
* Major reorganization to allow for other D20 games (D20Modern, etc) to
be used in concert. Also will now easily be able to split the Gemini
logic away from the Android UI, so the logic can be re-used in a PC
version of the chargen.  Added marker interfaces to identify things as
Fantasy, Modern, etc. and implemented throughout.
* Rewrote Library class to be much more dynamic/versatile, made it
non-static
* Changed D20SRD to use reflection to get the arrays of objects so we
don't have to declare every object twice.
* Beginning to work on logic for spellcasters
* Bug fixes
* Fleshing out more Classes
* Added hooks for psionics,prestige classes
* Finished up UI event listener structure