Split D20Character into GameCharacter
Pulled superclass GameCharacter out of D20Character and moved a few
other D20-specific classes, such that gemini.common is now clean of
anything d20-specific, in the hopes of eventually making the module work
with multiple game systems. Not bothering with trying to genericize
skills, etc. right now; this is too much work for virtually no benefit.
Because the listener architecture needs to know about characters (and
virtually all other screens are system-specific and can use the derived
type for the appropriate system), Characters seemed to be the only ones
that made sense to do right now (and probably ever).

On the plus side, it made D20Character *loads* easier to grok, as you're
not overwhelmed with stuff like "getEyeColor()".